+++ CHANGELOG.md
@@ -7,7 +7,11 @@ All notable changes to this project will be documented in this file. The version
 ### Added
 * New setting `autoBrakeAngle` to control at what angle offset from the intended path brakes automatically engage.
 * Press "G" (Lua binding "gear"), to engage auto parking.
+* Press `CTRL + SHIFT + G` to save the current position as a waypoint named "Tmp G".
+* Press `SHIFT + G` to automatically return to saved parking position.
 * New setting `parkMaxSpeed` to set maximum parking speed.
+* New setting `parkMargin` to control the margin of the parking route.
+* New setting `parkHeightMargin` to control how far above the park position to move before descending.
 
 ### Fixed
 * Corrected some broken math that had went unnoticed since pre v1.0.